Improvements to asynchronous MS writing
- Move the code into a module to work around
  https://bugs.python.org/issue9914
- Catch exceptions in ms_writer_process and pass them back to the
  caller.
